+++ MyGeotabAPIAdapter.Database/Models/DbDeviceStatusInfo.cs
@@ -31,7 +31,7 @@ namespace MyGeotabAPIAdapter.Database.Models
         public bool? IsHistoricLastDriver { get; set; }
         public double Latitude { get; set; }
         public double Longitude { get; set; }
-        public double Speed { get; set; }
+        public float Speed { get; set; }
         [ChangeTracker]
         public DateTime RecordLastChangedUtc { get; set; }
     }
